requirements.txt
aiogram
asyncio
loguru
psycopg
psycopg-binary
psycopg-pool
pydantic-settings
python-dotenv